backend/app/data_ingestion/scraper.py
# ============================================
# Boussole - Data Scraper Module
# ============================================

"""
Web scraping module for extracting data from various sources.

This module provides functions for scraping data from:
- ONS (Office National des Statistiques) website
- Government websites
- Other official data sources

Note: Always respect robots.txt and terms of service of target websites.
"""

import logging
from typing import Optional, List, Dict, Any
from bs4 import BeautifulSoup
import requests
import aiohttp
from urllib.parse import urljoin, urlparse

logger = logging.getLogger(__name__)


def scrape_data(
    source_url: Optional[str] = None,
    source_type: str = "ons",
    **kwargs
) -> List[Dict[str, Any]]:
    """
    Scrape data from a specified source.
    
    This is a template function. Add your scraping logic here.
    
    Args:
        source_url: URL to scrape from (optional, uses default if not provided)
        source_type: Type of source (ons, government, custom, etc.)
        **kwargs: Additional parameters for scraping
    
    Returns:
        List of dictionaries containing scraped data
    
    Example:
        >>> data = scrape_data("https://www.ons.dz", source_type="ons")
        >>> print(f"Scraped {len(data)} records")
    """
    logger.info(f"Starting data scraping from {source_type}: {source_url}")
    
    # ============================================
    # ADD SCRAPING LOGIC HERE
    # ============================================
    
    # Example: Simple HTTP request with BeautifulSoup
    try:
        if not source_url:
            logger.warning("No source URL provided, using default")
            source_url = "https://www.ons.dz"
        
        # Make HTTP request
        headers = {
            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
        }
        response = requests.get(source_url, headers=headers, timeout=30)
        response.raise_for_status()
        
        # Parse HTML content
        soup = BeautifulSoup(response.content, "html.parser")
        
        # Extract data based on HTML structure
        # This is where you add your specific scraping logic
        # Example:
        # data = []
        # for table in soup.find_all("table"):
        #     for row in table.find_all("tr"):
        #         cells = row.find_all("td")
        #         if len(cells) > 0:
        #             data.append({
        #                 "column1": cells[0].text.strip(),
        #                 "column2": cells[1].text.strip() if len(cells) > 1 else None,
        #             })
        
        logger.info(f"Scraping completed successfully")
        return []  # Return empty list as placeholder
    
    except requests.RequestException as e:
        logger.error(f"HTTP request failed: {e}")
        return []
    except Exception as e:
        logger.error(f"Scraping failed: {e}", exc_info=True)
        return []


async def scrape_async(
    source_url: str,
    source_type: str = "ons",
    **kwargs
) -> List[Dict[str, Any]]:
    """
    Asynchronously scrape data from a specified source.
    
    This is a template function for async scraping.
    Add your async scraping logic here.
    
    Args:
        source_url: URL to scrape from
        source_type: Type of source (ons, government, custom, etc.)
        **kwargs: Additional parameters for scraping
    
    Returns:
        List of dictionaries containing scraped data
    """
    logger.info(f"Starting async data scraping from {source_type}: {source_url}")
    
    # ============================================
    # ADD ASYNC SCRAPING LOGIC HERE
    # ============================================
    
    # Example: Async HTTP request with aiohttp
    try:
        async with aiohttp.ClientSession() as session:
            headers = {
                "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
            }
            async with session.get(source_url, headers=headers, timeout=30) as response:
                response.raise_for_status()
                html = await response.text()
                
                # Parse HTML content
                soup = BeautifulSoup(html, "html.parser")
                
                # Extract data based on HTML structure
                # Add your specific scraping logic here
                
                logger.info(f"Async scraping completed successfully")
                return []  # Return empty list as placeholder
    
    except aiohttp.ClientError as e:
        logger.error(f"Async HTTP request failed: {e}")
        return []
    except Exception as e:
        logger.error(f"Async scraping failed: {e}", exc_info=True)
        return []


def scrape_ons_data(
    category: Optional[str] = None,
    year: Optional[int] = None
) -> List[Dict[str, Any]]:
    """
    Scrape data specifically from ONS (Office National des Statistiques).
    
    Args:
        category: Data category to scrape (e.g., 'agriculture', 'energy')
        year: Specific year to scrape data for
    
    Returns:
        List of dictionaries containing scraped ONS data
    
    Note:
        This is a template. Implement actual ONS scraping logic here.
        Check https://www.ons.dz for available data.
    """
    logger.info(f"Scraping ONS data - Category: {category}, Year: {year}")
    
    # ============================================
    # ADD ONS-SPECIFIC SCRAPING LOGIC HERE
    # ============================================
    
    # Example implementation:
    # base_url = "https://www.ons.dz"
    # url = f"{base_url}/donnees/{category}" if category else base_url
    # 
    # # Add year filter if provided
    # if year:
    #     url += f"?annee={year}"
    # 
    # # Scrape data from the URL
    # data = scrape_data(url, source_type="ons")
    
    logger.info(f"ONS data scraping completed")
    return []  # Return empty list as placeholder


def scrape_government_data(
    ministry: str,
    data_type: str
) -> List[Dict[str, Any]]:
    """
    Scrape data from government ministry websites.
    
    Args:
        ministry: Ministry name or code (e.g., 'energy', 'agriculture')
        data_type: Type of data to scrape (e.g., 'statistics', 'reports')
    
    Returns:
        List of dictionaries containing scraped government data
    
    Note:
        This is a template. Implement actual government scraping logic here.
    """
    logger.info(f"Scraping government data - Ministry: {ministry}, Type: {data_type}")
    
    # ============================================
    # ADD GOVERNMENT-SPECIFIC SCRAPING LOGIC HERE
    # ============================================
    
    # Example implementation:
    # base_url = f"https://www.{ministry}.gov.dz"
    # url = f"{base_url}/{data_type}"
    # 
    # # Scrape data from the URL
    # data = scrape_data(url, source_type="government")
    
    logger.info(f"Government data scraping completed")
    return []  # Return empty list as placeholder


def check_robots_txt(url: str) -> bool:
    """
    Check if scraping is allowed by robots.txt.
    
    Args:
        url: URL to check
    
    Returns:
        True if scraping is allowed, False otherwise
    """
    try:
        parsed_url = urlparse(url)
        robots_url = f"{parsed_url.scheme}://{parsed_url.netloc}/robots.txt"
        
        response = requests.get(robots_url, timeout=10)
        if response.status_code == 200:
            # Parse robots.txt and check permissions
            # For simplicity, returning True here
            # In production, use robotparser module
            return True
        
        return True  # No robots.txt found, assume allowed
    
    except Exception as e:
        logger.warning(f"Failed to check robots.txt: {e}")
        return True  # Assume allowed if check fails
