backend/requirements.txt
# ============================================
# Boussole - Algerian Data Analytics SaaS
# Backend Dependencies (Python)
# ============================================

# ============================================
# Core Framework
# ============================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic
pydantic-settings
python-multipart==0.0.6

# ============================================
# Database
# ============================================
sqlalchemy
alembic==1.13.1
asyncpg
psycopg2-binary

# ============================================
# Authentication & Security
# ============================================
python-jose[cryptography]==3.3.0
passlib[bcrypt,argon2]==1.7.4
argon2-cffi
python-dotenv==1.0.0
pyotp==2.9.0

# ============================================
# Task Queue
# ============================================
celery==5.3.6
redis==5.0.1
flower==2.0.1

# ============================================
# AI / RAG / LLM
# ============================================
# langchain
# langchain-community
# langchain-openai
# langchain-groq
# sentence-transformers==2.3.1
# openai==1.10.0
# google-generativeai
google-generativeai
# tiktoken==0.5.2

# ============================================
# Data Processing
# ============================================
pandas
numpy>=2.1.0
openpyxl==3.1.2
openpyxl==3.1.2
xlrd==2.0.1
prophet

# ============================================
# Web Scraping
# ============================================
beautifulsoup4==4.12.3
requests==2.31.0
aiohttp
selenium==4.17.2
webdriver-manager==4.0.1

# ============================================
# HTTP Client
# ============================================
httpx==0.26.0

# ============================================
# Validation & Serialization
# ============================================
email-validator==2.1.0

# ============================================
# Utilities
# ============================================
python-dateutil==2.8.2
pytz==2023.3.post1
tenacity==8.2.3

# ============================================
# Testing
# ============================================
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# ============================================
# Code Quality
# ============================================
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2

# ============================================
# API Documentation
# ============================================
python-json-logger==2.0.7

# ============================================
# Monitoring & Logging
# ============================================
sentry-sdk==1.40.0
prometheus-client==0.19.0

# ============================================
# Payments
# ============================================
stripe


